dev-requirements.txt
-r requirements.txt

pytz==2015.7

django-behave==0.1.5
splinter==0.7.3
factory-boy==2.6.0
